Delete poster and rename poster2.
