+++ main.py
@@ -1,6 +1,6 @@
 import logging
 
-from project_den import figure, poster2, util
+from project_den import figure, poster, util
 
 
 logging.basicConfig(level=logging.INFO)
@@ -31,6 +31,6 @@ for fig_name, fig_config in configs['figures'].items():
 # Combine figures into each poster.
 logging.info("Building poster from figures and layout...")
 for poster_name, poster_kwargs in configs['posters'].items():
-    poster = poster2.build_poster(figures=figures, **poster_kwargs)
+    poster = poster.build_poster(figures=figures, **poster_kwargs)
     logging.debug(f"Poster created: {poster_name}")
